# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/books/free": {
      "get": {
        "operationId": "BooksController_getFreeBooks",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "maxResults",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Books"
        ]
      }
    },
    "/api/books/category/{category}": {
      "get": {
        "operationId": "BooksController_getByCategory",
        "parameters": [
          {
            "name": "category",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "maxResults",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Books"
        ]
      }
    },
    "/api/books/{id}": {
      "get": {
        "operationId": "BooksController_getBookById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Books"
        ]
      }
    }
  },
  "info": {
    "title": "Readify API",
    "description": "The API documentation for Readify backend application",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "readify_token",
        "description": "Session cookie for authenticating requests"
      }
    },
...
  }
}